Xception variant model up and running. ~55% validation accuracy on 5 species after 200 epochs Epoch 200/200 28/28 [==============================] - 28s 981ms/step - loss: 0.1379 - accuracy: 0.9516 - val_loss: 2.9326 - val_accuracy: 0.5568
Still need to run Keras_tuner on Xception model as well as figure out issue with batch normalization and if it should be included in the model. This model is serving as a foundation, I will still need to research each of its components to get a better understanding of the model and what I can do to fine tune it for my project. The training accuracy is much higher than the validation accuracy, which stagnates pretty quickly.